docs: update docs/scripts for installing Poetry and its dependencies

+++ bin/install_poetry.sh
@@ -5,6 +5,17 @@ set -euo pipefail
 export DEBIAN_FRONTEND=noninteractive
 
 function install_apt_packages() {
+    if [[ "$OSTYPE" == 'darwin'* ]]; then
+        echo "Skipping installation of APT packages. Detected macOS."
+        return
+    fi
+
+    if [[ "$OSTYPE" == 'linux'* && ! -f /etc/debian_version ]]; then
+        echo "Skipping installation of APT packages. Detected non-Debian-based Linux distribution."
+        return
+    fi
+
+    sudo apt-get update -y && \
     sudo apt-get install -y \
         apt-utils \
         curl \
@@ -14,10 +25,21 @@ function install_apt_packages() {
 }
 
 function install_python() {
-    sudo -E apt-get install -y \
-        python3 \
-        python3-pip \
-        python-is-python3
+    if [[ "$OSTYPE" == 'darwin'* ]]; then
+        # only install python3.11 on macOS if it's not already installed
+        if brew list | grep -q python@3.11; then
+            echo "Python 3.11 already installed. Skipping installation."
+        else
+            brew install python@3.11
+        fi
+    elif [[ "$OSTYPE" == 'linux'* && -f /etc/debian_version ]]; then
+        sudo -E apt-get install -y \
+            python3 \
+            python3-pip \
+            python-is-python3
+    else
+        echo "Skipping installation of Python. OS not supported."
+    fi
 }
 
 function install_pyenv {
@@ -40,7 +62,6 @@ function install_poetry() {
     curl -sSL https://install.python-poetry.org | python3 -
 }
 
-sudo apt-get update -y
 install_apt_packages
 install_python
 install_pyenv
